fix(desktop): honor in-app subtitle style + shadow on the embed player
The embed (Windows) player set only font size, colour, background and position,
at mpv's oversized default base — it ignored sub-ass-override, the edge type
and the outline/shadow, so subtitles dropped the configured drop-shadow and let
embedded ASS styling through. Extract the compositor's full sub-* mapper into a
shared module and drive both backends from it, so the embed player renders the
same calibrated style + shadow as the compositor and they can't drift again.
